requirements.txt
discord.py[voice]
PyNaCl
youtube_dl
requests
